refactor: @Repository 어노테이션 추가 및 enums 폴더 생성
