+++ src/main/java/com/study/platform/blog/domain/Blog.java
@@ -16,9 +16,9 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
+@Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Getter
-@Entity
 public class Blog {
 
     @Id
@@ -42,3 +42,4 @@ public class Blog {
         this.title = title;
     }
 }
+
