+++ src/main/java/com/study/platform/blog/domain/BlogRepository.java
@@ -1,6 +1,8 @@
 package com.study.platform.blog.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface BlogRepository extends JpaRepository<Blog, Long> {
 }
