--- src/main/java/com/study/platform/member/domain/Authority.java
@@ -1,5 +1,0 @@
-package com.study.platform.member.domain;
-
-public enum Authority {
-    ADMIN,MEMBER
-}
