+++ src/main/java/com/study/platform/member/domain/MemberRepository.java
@@ -1,6 +1,8 @@
 package com.study.platform.member.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
 }
