+++ src/main/java/com/study/platform/member/domain/enums/Authority.java
@@ -1,0 +1,5 @@
+package com.study.platform.member.domain.enums;
+
+public enum Authority {
+    ADMIN,MEMBER
+}
