+++ src/main/java/com/study/platform/member/domain/enums/DeveloperType.java
@@ -1,4 +1,4 @@
-package com.study.platform.member.domain;
+package com.study.platform.member.domain.enums;
 
 public enum DeveloperType {
     FRONTEND,BACKEND
